+++ tests/unit/core/test_views.py
@@ -7,7 +7,6 @@ from directory_api_client import api_client
 from formtools.wizard.views import normalize_name
 import pytest
 
-from django.db.utils import DataError
 from django.urls import reverse
 
 from core import forms, helpers, models, serializers, views, constants
@@ -172,6 +171,7 @@ def test_dashboard_page_lesson_progress(
     mock_get_company_profile,
     client,
     user,
+    get_request,
     domestic_homepage,
     domestic_site
 ):
@@ -199,7 +199,7 @@ def test_dashboard_page_lesson_progress(
     )
     # create dashboard
     dashboard = DomesticDashboardFactory(parent=domestic_homepage, slug='dashboard')
-    context_data = dashboard.get_user_context(user)
+    context_data = dashboard.get_context(get_request)
     # check the progress
     assert len(context_data['list_pages']) == 2
     assert context_data['list_pages'][0] == topic_one
@@ -211,37 +211,10 @@ def test_dashboard_page_lesson_progress(
 
 
 @pytest.mark.django_db
-@mock.patch.object(api_client.personalisation, 'events_by_location_list')
-@mock.patch.object(api_client.personalisation, 'export_opportunities_by_relevance_list')
-def test_dashboard_page_lesson_division_by_zero(
-        mock_events_by_location_list,
-        mock_export_opportunities_by_relevance_list,
-        patch_set_user_page_view,
-        patch_get_user_page_views,
-        mock_get_company_profile,
-        client,
-        user,
-        domestic_homepage,
-        domestic_site
-):
-    mock_events_by_location_list.return_value = create_response(json_body={'results': []})
-    mock_export_opportunities_by_relevance_list.return_value = create_response(json_body={'results': []})
-
-    # given a lesson listing page without any lesson in it
-    ListPageFactory(parent=domestic_homepage, slug='test-topic-one', record_read_progress=True)
-
-    # when the dashboard is visited a division by zero should be raised
-    with pytest.raises(DataError, match='division by zero'):
-        # create dashboard
-        client.force_login(user)
-        dashboard = DomesticDashboardFactory(parent=domestic_homepage, slug='dashboard')
-        dashboard.get_user_context(user)
-
-
-@pytest.mark.django_db
 def test_dashboard_apis_ok(
     client,
     user,
+    get_request,
     patch_get_dashboard_events,
     patch_get_dashboard_export_opportunities,
     patch_set_user_page_view,
@@ -288,7 +261,7 @@ export-opportunities/opportunities/french-sardines-required',
             client.force_login(user)
 
             dashboard = DomesticDashboardFactory(parent=domestic_homepage, slug='dashboard')
-            context_data = dashboard.get_user_context(user)
+            context_data = dashboard.get_context(get_request)
 
             assert context_data['events'] == [{
                 'title': 'Global Aid and Development Directory',
@@ -318,6 +291,7 @@ export-opportunities/opportunities/french-sardines-required',
 def test_dashboard_apis_fail(
         client,
         user,
+        get_request,
         patch_get_dashboard_events,
         patch_get_dashboard_export_opportunities,
         patch_set_user_page_view,
@@ -339,7 +313,7 @@ personalisation.export_opportunities_by_relevance_list'
 
             client.force_login(user)
             dashboard = DomesticDashboardFactory(parent=domestic_homepage, slug='dashboard')
-            context_data = dashboard.get_user_context(user)
+            context_data = dashboard.get_context(get_request)
 
             assert context_data['events'] == []
             assert context_data['export_opportunities'] == []